+++ src/seismometer/api/__init__.py
@@ -1,6 +1,17 @@
+# flake8: noqa: F405 -- allow wildcard imports
+
 from seismometer.configuration import ConfigProvider
 from seismometer.core.logger import init_logger
 from seismometer.data.loader import loader_factory
 from seismometer.seismogram import Seismogram
 
+from .explore import *
+from .plots import *
+from .reports import *
+from .templates import *
+
 __all__ = ["ConfigProvider", "loader_factory", "Seismogram", "init_logger"]
+__all__.extend(explore.__all__)
+__all__.extend(reports.__all__)
+__all__.extend(plots.__all__)
+__all__.extend(templates.__all__)
